+++ backend/src/test/java/story/cheek/common/annotation/RepositoryTest.java
@@ -1,0 +1,21 @@
+package story.cheek.common.annotation;
+
+import static org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace.NONE;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
+import story.cheek.DatabaseCleanup;
+import story.cheek.config.TestQueryDslConfig;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+@AutoConfigureTestDatabase(replace = NONE)
+@DataJpaTest
+@Import({TestQueryDslConfig.class, DatabaseCleanup.class})
+public @interface RepositoryTest {
+}
